+++ crates/ruma-state-res/src/error.rs
@@ -1,4 +1,4 @@
-use serde_json::Error as JsonError;
+use ruma_common::{OwnedEventId, RoomVersionId};
 use thiserror::Error;
 
 /// Result type for state resolution.
@@ -8,32 +8,19 @@ pub type Result<T> = std::result::Result<T, Error>;
 #[derive(Error, Debug)]
 #[non_exhaustive]
 pub enum Error {
-    /// A deserialization error.
-    #[error(transparent)]
-    SerdeJson(#[from] JsonError),
-
     /// The given option or version is unsupported.
     #[error("Unsupported room version: {0}")]
-    Unsupported(String),
+    UnsupportedRoomVersion(RoomVersionId),
 
     /// The given event was not found.
-    #[error("Not found error: {0}")]
-    NotFound(String),
+    #[error("Failed to find event {0}")]
+    NotFound(OwnedEventId),
 
-    /// Invalid fields in the given PDU.
-    #[error("Invalid PDU: {0}")]
-    InvalidPdu(String),
+    /// An auth event is invalid.
+    #[error("Invalid auth event: {0}")]
+    AuthEvent(String),
 
-    /// A custom error.
-    #[error("{0}")]
-    Custom(Box<dyn std::error::Error>),
-}
-
-impl Error {
-    pub fn custom<E>(e: E) -> Self
-    where
-        E: Into<Box<dyn std::error::Error>>,
-    {
-        Self::Custom(e.into())
-    }
+    /// A state event doesn't have a `state_key`.
+    #[error("State event has no `state_key`")]
+    MissingStateKey,
 }